requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
yt_dlp
aiofiles